refactor(components/formfieldmedia): adds a new component that composes modal and input
