+++ src/components/FormFieldMedia/index.js
@@ -1,0 +1,3 @@
+import FormFieldMedia from "./FormFieldMedia"
+
+export default FormFieldMedia
